+++ source/Energinet.DataHub.MarketParticipant.Application/Validation/GetUserRoleAssignmentAuditLogEntriesCommandRuleSet.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.MarketParticipant.Application.Commands.UserRoleTemplates;
+using Energinet.DataHub.MarketParticipant.Application.Commands.UserRoles;
 using FluentValidation;
 
 namespace Energinet.DataHub.MarketParticipant.Application.Validation