+++ source/Energinet.DataHub.MarketParticipant.Client/source/Energinet.DataHub.MarketParticipant.Client/IMarketParticipantUserRoleAssignmentClient.cs
@@ -15,21 +15,20 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Energinet.DataHub.MarketParticipant.Client.Models;
 
 namespace Energinet.DataHub.MarketParticipant.Client
 {
     /// <summary>
-    /// Manages user role templates.
+    /// Manages user roles.
     /// </summary>
     public interface IMarketParticipantUserRoleAssignmentClient
     {
         /// <summary>
-        /// Updates all user roles assignments
+        /// Updates user roles assignments for the specified user and actor.
         /// </summary>
+        /// <param name="actorId">The id of the actor.</param>
         /// <param name="userId">The id of the user.</param>
-        /// <param name="updateUserRoleAssignmentsDto">user role updates.</param>
-        /// <returns>The list of assignable user role templates.</returns>
-        Task UpdateUserRoleAssignmentsAsync(Guid userId, UpdateUserRoleAssignmentsDto updateUserRoleAssignmentsDto);
+        /// <param name="userRoleAssignments">The new user roles.</param>
+        Task UpdateUserRoleAssignmentsAsync(Guid actorId, Guid userId, IEnumerable<Guid> userRoleAssignments);
     }
 }