+++ source/Energinet.DataHub.MarketParticipant.Domain/Model/Users/UserOverviewItem.cs
@@ -16,4 +16,10 @@ using System;
 
 namespace Energinet.DataHub.MarketParticipant.Domain.Model.Users;
 
-public sealed record UserOverviewItem(UserId Id, EmailAddress Email, string Name, string? PhoneNumber, DateTimeOffset CreatedDate, bool Active);
+public sealed record UserOverviewItem(
+    UserId Id,
+    EmailAddress Email,
+    string Name,
+    PhoneNumber? PhoneNumber,
+    DateTimeOffset CreatedDate,
+    bool Active);