+++ source/Energinet.DataHub.MarketParticipant.Libraries.Tests/Clients/MarketParticipantClientTests.cs
@@ -878,24 +878,24 @@ public sealed class MarketParticipantClientTests
         // arrange
         var id = Guid.Parse("8CC33D1F-2E3D-4752-A231-851A2C0589D9");
         const string gln = "9326018586901";
-        const string name = "Actor";
+        const string name = "Organization name";
 
         var incomingJson = $@"
             [
               {{
                 ""id"": ""{id}"",
                 ""gln"": ""{gln}"",
-                ""name"": ""{name}""
+                ""organizationName"": ""{name}""
               }},
               {{
                 ""id"": ""{id}"",
                 ""gln"": ""{gln}"",
-                ""name"": ""{name}""
+                ""organizationName"": ""{name}""
               }},
               {{
                 ""id"": ""{id}"",
                 ""gln"": ""{gln}"",
-                ""name"": ""{name}""
+                ""organizationName"": ""{name}""
               }},
             ]";